% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_helpers.R
\name{add_athlete_id_col}
\alias{add_athlete_id_col}
\title{Add Athlete ID's to data frame}
\usage{
add_athlete_id_col(df, name_col, team_col = NULL, headshot_urls = FALSE)
}
\arguments{
\item{df}{a data frame.}

\item{name_col}{the column in \code{df} with the player names to join with the roster data.}

\item{team_col}{Optional column with team names to join with the roster data to reduce
the chance of matching with two players from different teams with the same name. If NULL
and \code{df} has a column named "team" or "school," the function will use those as \code{team_col}.
The function also checks \code{df} for a column named "season" to match names to rosters going
back to 2009.}

\item{headshot_urls}{logical to return headshot URLs. If TRUE, the output has an additional
column called "headshot_url" with links for player headshots.}
}
\value{
the original \code{df} with extra columns:
\describe{
\item{\code{athlete_id}}{athlete ESPN ID.}
\item{\code{headshot_url}}{url of athlete's headshot.}
}
}
\description{
This function attempts to add ESPN athlete ID's to a data frame using the roster data
in the cfbfastR-data repo. The function is experimental and not guaranteed to be accurate.
}
\examples{
\donttest{
x <- data.frame(
  player_name = c("Britain Covey","JT Daniels")
)

cfbplotR::add_athlete_id_col(x, player_name)


x$season <- c(2021,2021)
cfbplotR::add_athlete_id_col(x, player_name)


x$team = c("Utah","Georgia")
cfbplotR::add_athlete_id_col(x, player_name, team, headshot_urls = TRUE)
}
}
