% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_helpers.R
\name{clean_school_names}
\alias{clean_school_names}
\title{Standardize NCAA School Names}
\usage{
clean_school_names(school, keep_non_matches = TRUE)
}
\arguments{
\item{school}{a character vector of names}

\item{keep_non_matches}{If \code{TRUE} (the default) an element of \code{school} that can't
be matched to any of the internal mapping vectors will be kept as is. Otherwise
it will be replaced with \code{NA}.}
}
\value{
A character vector with the length of \code{school} and cleaned team abbreviations
if they are included in \code{\link{team_name_mapping}}. Non matches may be replaced
with \code{NA} (depending on the value of \code{keep_non_matches}).
}
\description{
This function standardizes college names to cfbplotR defaults.
This helps for joins and plotting.
}
\examples{

x <- c("utah", "San Jose State", "Hawaii", "UTSA", "SLC", "USC")
# use current location and keep non matches
clean_school_names(x)

# replace non matches
clean_school_names(x, keep_non_matches = FALSE)
}
