% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_helpers.R
\name{clean_team_abbrs}
\alias{clean_team_abbrs}
\title{Clean CFB team abbreviations}
\usage{
clean_team_abbrs(school, keep_non_matches = TRUE)
}
\arguments{
\item{school}{a character vector of names.}

\item{keep_non_matches}{If \code{TRUE} (the default) an element of \code{school} that
can't be matched to any of the internal mapping vectors will be kept as is.
Otherwise it will be replaced with \code{NA}.}
}
\value{
A character vector of cleaned team names.
}
\description{
Alias of \code{\link[=clean_school_names]{clean_school_names()}} for naming parity with the
nflplotR/nbaplotR family.
}
