% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{logo_ref}
\alias{team_name_mapping}
\title{\strong{Data in the package for reference}}
\format{
A data frame with 830+ rows and 6 variables:
\describe{
\item{school}{School name}
\item{type}{Conference-tier (i.e. P5, G5, etc.)}
\item{logo}{primary school logo from ESPN.com}
\item{color}{current primary school color}
\item{alt_color}{current secondary school color}
\item{wordmark}{Wordmark for school if available}
}

A vector with 1100+ name variations
\describe{
\item{Short Name}{to Full Naame}
}
}
\usage{
logo_ref

team_name_mapping
}
\description{
A dataset containing the school names, colors & logos
for all college teams (NCAA DI/DII/DIII and FBS/FCS).

A dataset mapping from abbreviation or alternate name forms to most commonly used.
}
\keyword{data}
