% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtitle_image.R
\name{ggtitle_image}
\alias{ggtitle_image}
\alias{theme_title_image}
\title{Functions for adding an image to the title of a ggplot}
\usage{
ggtitle_image(
  title_image = waiver(),
  title = waiver(),
  image_height = 15,
  image_side = c("left", "right"),
  subtitle = waiver()
)

theme_title_image(...)
}
\arguments{
\item{title_image}{The URL of the image to add to the title. If one of \code{\link[=valid_team_names]{valid_team_names()}}, the school logo will be used automatically.}

\item{title}{The text for the title. The title has to be defined here as the function overwrites any current title text.}

\item{image_height}{The height of the image in pixels.}

\item{image_side}{One of "left" or "right". Places the image on either side of the title text.}

\item{subtitle}{Optional text for the subtitle. Does not overwrite current subtitle if left blank.}

\item{...}{Other arguments passed on to \code{\link[ggtext:element_markdown]{ggtext::element_markdown()}}.}
}
\description{
These functions work together to place an image to the left or right of the title in a ggplot. \code{\link{ggtitle_image}} is the main function but must be used with either theme_title_image() or setting the plot.title argument in \code{\link[ggplot2:theme]{ggplot2::theme()}} to \code{\link[ggtext:element_markdown]{ggtext::element_markdown()}}
}
\examples{
\donttest{
library(cfbplotR)
library(ggplot2)

logo_url <- "https://github.com/Kazink36/cfbplotR/raw/master/man/figures/logo.png"

  p <- ggplot(mtcars, aes(x = hp, y = mpg)) +
    geom_point() +
    labs(title = "This Title will be overwritten",
         subtitle = "This is the Subtitle")

if (utils::packageVersion("gridtext") > "0.1.4"){
  p +
    ggtitle_image(title_image = logo_url,
                  title = "This is the Title",
                  image_height = 40,
                  image_side = "right") +
    theme_title_image(size = 20, hjust = 0.5)
}
if (utils::packageVersion("gridtext") > "0.1.4"){
  p +
    ggtitle_image(title_image = "Utah",
                  title = "This it the Title",
                  image_height = 20,
                  image_side = "left",
                  subtitle = "This overwrites the old subtitle") +
    theme(plot.title = ggtext::element_markdown(size = 20, hjust = 0.5))
}
}

}
\seealso{
\code{\link[=ggtitle_image]{ggtitle_image()}}, \code{\link[=theme_title_image]{theme_title_image()}}
}
