% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle_geoms.R
\name{median_geom}
\alias{median_geom}
\title{Plot Middle Lines}
\usage{
median_geom(x, y, color = "red", linetype = "dashed")
}
\arguments{
\item{x}{The X data}

\item{y}{The Y data}

\item{color}{The color of the line}

\item{linetype}{The linetype of the line}
}
\description{
These functions plot lines representing the median or mean of the data for ggplot. From Thomas Mock (https://github.com/nflverse/nflplotR/issues/1)
}
\examples{
\donttest{
ggplot(mtcars) +
 geom_point(aes(x = disp, y = mpg)) +
 median_geom(disp, mpg)


ggplot(mtcars) +
 geom_point(aes(x = disp, y = mpg)) +
 mean_geom(disp, mpg)
 }
}
