% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geom_from_path}
\alias{GeomFromPath}
\alias{element_path}
\alias{element_raster}
\alias{geom_mean_lines}
\alias{geom_median_lines}
\alias{GeomRefLines}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggpath}{\code{\link[ggpath]{element_path}}, \code{\link[ggpath:element_path]{element_raster}}, \code{\link[ggpath]{geom_from_path}}, \code{\link[ggpath:geom_lines]{geom_mean_lines}}, \code{\link[ggpath:geom_lines]{geom_median_lines}}, \code{\link[ggpath:ggpath-package]{GeomFromPath}}, \code{\link[ggpath:ggpath-package]{GeomRefLines}}}
}}

