% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cfb.R
\name{scale_color_cfb}
\alias{scale_color_cfb}
\alias{scale_colour_cfb}
\alias{scale_fill_cfb}
\title{Scale for college football team colors}
\usage{
scale_color_cfb(
  alt_colors = NULL,
  values = NULL,
  ...,
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)

scale_colour_cfb(
  alt_colors = NULL,
  values = NULL,
  ...,
  aesthetics = "colour",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)

scale_fill_cfb(
  alt_colors = NULL,
  values = NULL,
  ...,
  aesthetics = "fill",
  breaks = ggplot2::waiver(),
  na.value = "grey50",
  guide = NULL,
  alpha = NA
)
}
\arguments{
\item{alt_colors}{Vector of team names to use the alternate color of.}

\item{...}{Arguments passed on to scale_color_manual}
}
\description{
These functions allows you to map college football team names as levels to the color and fill aesthetics
}
\examples{
library(cfbplotR)
library(ggplot2)

df <- data.frame(
  y = 6:9,
  teams = c("Alabama","Florida State","Oregon","Utah")
 )
 ggplot(df, aes(x = teams, y = y)) +
   geom_col(aes(color = teams, fill = teams), size = 2) +
   scale_color_cfb(alt_colors = df$teams) +
   scale_fill_cfb() +
   theme_minimal()
}
