% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cfb.R
\name{theme_cfb}
\alias{theme_x_cfb}
\alias{theme_y_cfb}
\title{Theme for CFB Team Logos}
\usage{
theme_x_cfb()

theme_y_cfb()
}
\description{
These functions are convenience wrappers around a theme call
that activates markdown in x-axis and y-axis labels
using \code{\link[ggtext:element_markdown]{ggtext::element_markdown()}}.
}
\details{
These functions are a wrapper around the function calls
\code{ggplot2::theme(axis.text.x = ggtext::element_markdown())} as well as
\code{ggplot2::theme(axis.text.y = ggtext::element_markdown())}.
They are made to be used in conjunction with \code{\link[=scale_x_cfb]{scale_x_cfb()}} and
\code{\link[=scale_y_cfb]{scale_y_cfb()}} respectively.
}
\examples{
library(cfbplotR)
library(ggplot2)

team_abbr <- cfbplotR::valid_team_names()
# remove conference logos from this example
team_abbr <- team_abbr[1:16]

df <- data.frame(
  random_value = runif(length(team_abbr), 0, 1),
  teams = team_abbr
)

if (utils::packageVersion("gridtext") > "0.1.4"){
  ggplot(df, aes(x = teams, y = random_value)) +
    geom_col(aes(color = teams, fill = teams), width = 0.5) +
    scale_color_cfb(alt_colors = team_abbr) +
    scale_fill_cfb(alpha = 0.4) +
    scale_x_cfb() +
    theme_minimal() +
    # theme_*_cfb requires gridtext version > 0.1.4
    theme_x_cfb()
}
}
\seealso{
\code{\link[=theme_x_cfb]{theme_x_cfb()}}, \code{\link[=theme_y_cfb]{theme_y_cfb()}}
}
