% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid CFB Team Names and Abbreviations}
\usage{
valid_team_names()
}
\description{
Output Valid CFB Team Names and Abbreviations
}
\examples{
# List valid team abbreviations excluding duplicates
valid_team_names()
}
