% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfbplotR-package.R, R/geom_cfb_wordmarks.R,
%   R/geom_from_path.R, R/geom_lines.R
\docType{package}
\name{cfbplotR-package}
\alias{cfbplotR}
\alias{cfbplotR-package}
\alias{GeomCFBwordmark}
\alias{GeomFromPath}
\alias{GeomRefLines}
\title{cfbplotR: CFB Logo Plots in 'ggplot2'}
\format{
An object of class \code{GeomCFBwordmark} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomFromPath} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomRefLines} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomCFBwordmark

GeomFromPath

GeomRefLines
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions to visualize College Football analysis in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cfbplotR.sportsdataverse.org}
  \item \url{https://github.com/sportsdataverse/cfbplotR}
  \item Report bugs at \url{https://github.com/sportsdataverse/cfbplotR/issues}
}

}
\author{
\strong{Maintainer}: Jared Lee \email{13jaredlee@gmail.com}

Authors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com}
}

Other contributors:
\itemize{
  \item Saiem Gilani \email{saiem.gilani@gmail.com} [contributor]
}

}
\keyword{datasets}
\keyword{internal}
