% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid CFB Team Names and Abbreviations}
\usage{
valid_team_names(
  division = c("FBS", "P5", "G5", "FCS", "DII", "DIII", "Conference", "hoopR", "Other")
)
}
\description{
Output Valid CFB Team Names and Abbreviations
}
\examples{
# List valid team abbreviations excluding duplicates
cfbplotR::valid_team_names("FBS")
cfbplotR::valid_team_names("FCS")
cfbplotR::valid_team_names("DII")
cfbplotR::valid_team_names("DIII")
cfbplotR::valid_team_names("Conference")
}
