% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid CFB Team Names and Abbreviations}
\usage{
valid_team_names(division = c("FBS", "FCS", "DII", "DIII", "Conference"))
}
\description{
Output Valid CFB Team Names and Abbreviations
}
\examples{
# List valid team abbreviations excluding duplicates
valid_team_names("FBS")
valid_team_names("FCS")
valid_team_names("DII")
valid_team_names("DIII")
valid_team_names("Conference")
}
